+++ MSTest/assert_utility.h
@@ -5,5 +5,5 @@ void assertAreEqual(
 	const std::vector<std::vector<double>> &expected,
 	const std::vector<std::vector<double>> &actual);
 void assertAreEqual(
-	const std::vector<double> &expected,
-	const std::vector<double> &actual);
+	const std::vector<size_t> &expected,
+	const std::vector<size_t> &actual);
